Итог работы 13 спринта

+++ controllers/cards.js
@@ -1,9 +1,10 @@
-const Card = require("../models/card");
+const mongoose = require('mongoose');
+const Card = require('../models/card');
 
 const getCards = (req, res) => {
   Card.find()
     .then((cards) => res.status(200).send(cards))
-    .catch(() => res.status(500).send({ message: err.message }));
+    .catch((err) => res.status(500).send({ message: err.message }));
 };
 
 const createCard = (req, res) => {
@@ -12,9 +13,9 @@ const createCard = (req, res) => {
   return Card.create(cardData)
     .then((card) => res.status(201).send(card))
     .catch((err) => {
-      if (err.name === "ValidationError") {
+      if (err.name === 'ValidationError') {
         return res.status(400).send({
-          message: "Переданы некорректные данные при создании карточки",
+          message: 'Переданы некорректные данные при создании карточки',
         });
       }
       return res.status(500).send({ message: err.message });
@@ -22,36 +23,39 @@ const createCard = (req, res) => {
 };
 
 const deleteCard = (req, res) => {
-  Card.findByIdAndDelete(req.params.cardId)
+  if (!mongoose.Types.ObjectId.isValid(req.params.cardId)) {
+    return res.status(400).send({ message: 'Некорректный _id карточки' });
+  }
+  return Card.findByIdAndDelete(req.params.cardId)
     .then((card) => {
       if (!card) {
         return res
           .status(404)
-          .send({ message: "Карточка с указанным _id не найдена" });
+          .send({ message: 'Карточка с указанным _id не найдена' });
       }
-      res.send({ message: "Карточка удалена" });
+      return res.send({ message: 'Карточка удалена' });
     })
-    .catch(() => res.status(500).send({ message: err.message }));
+    .catch((err) => res.status(500).send({ message: err.message }));
 };
 
 const likeCard = (req, res) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
-    { new: true }
+    { new: true },
   )
     .then((card) => {
       if (!card) {
         return res.status(404).send({
-          message: "Передан несуществующий _id карточки",
+          message: 'Передан несуществующий _id карточки',
         });
       }
       return res.status(200).send(card);
     })
     .catch((err) => {
-      if (err.name === "CastError") {
+      if (err.name === 'CastError') {
         return res.status(400).send({
-          message: "Переданы некорректные данные для постановки лайка",
+          message: 'Переданы некорректные данные для постановки лайка',
         });
       }
       return res.status(500).send({ message: err.message });
@@ -62,21 +66,21 @@ const dislikeCard = (req, res) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
-    { new: true }
+    { new: true },
   )
     .then((card) => {
       if (!card) {
         return res.status(404).send({
-          message: "Передан несуществующий _id карточки",
+          message: 'Передан несуществующий _id карточки',
         });
       }
       res.status(200).send(card);
       return card;
     })
     .catch((err) => {
-      if (err.name === "CastError") {
+      if (err.name === 'CastError') {
         return res.status(400).send({
-          message: "Переданы некорректные данные для постановки/снятии лайка",
+          message: 'Переданы некорректные данные для постановки/снятии лайка',
         });
       }
       return res.status(500).send({ message: err.message });
